Return @mentions preceeded by @
Such as @@kn
These changes bring the library into alignment with the actual Twitter
website and API as shown here:
https://twitter.com/StreetGeekEnt/statuses/554800426768277504
https://twitter.com/kung_fu_mike/status/570024458640973824

This commit only fixes the Ruby, Java, and JavaScript Libraries. I
don't have an environment for Objective C development.

A special thank you to @dustyjewett for his help setting up and running
Nodejs and Grunt for the JS portion of this commit.

This commit was done on the dime of Simply Measured
https://github.com/simplymeasured/
https://twitter.com/simplymeasured